+++ resources/views/components/questions-section.blade.php
@@ -9,7 +9,12 @@
                     {{--    Full name, img date     --}}
                     <div class="name-img d-flex justify-between">
                         <div class="d-flex align-center">
-                            <img style="border-radius: 30% !important;" src="{{ asset($question->user->image_sizes['100x100'] ?? $question->user->avatar)}}" alt="{{'Савлдиҳанда '. $question->user->fullName() }}" class="mr-3">
+                            <img style="border-radius: 30% !important;" 
+                                 src="{{ asset($question->user->image_sizes['100x100'] ?? $question->user->avatar)}}" 
+                                 alt="{{'Савлдиҳанда '. $question->user->fullName() }}" 
+                                 class="mr-3"
+                                 width="40"
+                                 >
                             <div class="name">
                                 <h4 class="mb-3">
                                     <a href="{{ route('users.show', $question->user->id)}}">